Shared validator and new checks
