+++ application/routers/fact.py
@@ -31,6 +31,9 @@ from application.core.utils import (
     DigitalLandJSONResponse,
 )
 
+from application.search.validators import validate_dataset
+from application.data_access.dataset_queries import get_dataset_names
+
 router = APIRouter()
 logger = logging.getLogger(__name__)
 
@@ -133,19 +136,14 @@ def get_fact(
         fact_dict = _convert_resources_to_dict(fact_dict)
         dataset_obj = get_dataset_query(session, query_params["dataset"])
         if dataset_obj is None:
-            logger.warning(
-                "Dataset metadata not found for %s; using dataset slug as fallback",
-                query_params["dataset"],
-            )
-        # Fallback to the dataset slug if metadata cannot be found
-        dataset_name = dataset_obj.name if dataset_obj else query_params["dataset"]
+            raise HTTPException(status_code=404, detail="dataset not found")
         return templates.TemplateResponse(
             request,
             "fact.html",
             {
                 "fact": fact_dict,
                 "pipeline_name": query_params["dataset"],
-                "dataset": {"name": dataset_name},
+                "dataset": {"name": dataset_obj.name},
                 "references": [],
                 "breadcrumb": [],
                 "schema": None,
@@ -262,3 +260,44 @@ router.add_api_route(
     response_class=HTMLResponse,
     include_in_schema=False,
 )
+
+
+def validate_fact_dataset_query_filters(
+    request: Request,
+    session: Session = Depends(get_session),
+) -> FactDatasetQueryFilters:
+    dataset = request.query_params.get("dataset")
+    if not dataset:
+        raise RequestValidationError(
+            errors=[
+                {
+                    "loc": ["query", "dataset"],
+                    "msg": "Missing required 'dataset' query parameter.",
+                    "type": "value_error",
+                }
+            ]
+        )
+    # datasette still holds sqlite for retired datasets, postgres is the source of truth
+    validate_dataset([dataset], get_dataset_names(session))
+    return FactDatasetQueryFilters(dataset=dataset)
+
+
+def validate_fact_query_filters(
+    request: Request,
+    session: Session = Depends(get_session),
+) -> FactQueryFilters:
+    dataset = request.query_params.get("dataset")
+    entity = request.query_params.get("entity")
+    if not (dataset and entity and entity.isdigit() and int(entity) >= 1):
+        raise RequestValidationError(
+            errors=[
+                {
+                    "loc": ["query"],
+                    "msg": "Invalid or missing query parameters.",
+                    "type": "value_error",
+                }
+            ]
+        )
+    validate_dataset([dataset], get_dataset_names(session))
+    field = request.query_params.getlist("field") or None
+    return FactQueryFilters(dataset=dataset, entity=int(entity), field=field)
